+++ compose.yml
@@ -59,7 +59,6 @@ services:
       - node_modules_ghost_i18n:/home/ghost/ghost/i18n/node_modules:delegated
       - node_modules_ghost_identity-token-service:/home/ghost/ghost/identity-token-service/node_modules:delegated
       - node_modules_ghost_importer-handler-content-files:/home/ghost/ghost/importer-handler-content-files/node_modules:delegated
-      - node_modules_ghost_importer-revue:/home/ghost/ghost/importer-revue/node_modules:delegated
       - node_modules_ghost_in-memory-repository:/home/ghost/ghost/in-memory-repository/node_modules:delegated
       - node_modules_ghost_job-manager:/home/ghost/ghost/job-manager/node_modules:delegated
       - node_modules_ghost_link-redirects:/home/ghost/ghost/link-redirects/node_modules:delegated
@@ -219,7 +218,6 @@ volumes:
   node_modules_ghost_i18n: {}
   node_modules_ghost_identity-token-service: {}
   node_modules_ghost_importer-handler-content-files: {}
-  node_modules_ghost_importer-revue: {}
   node_modules_ghost_in-memory-repository: {}
   node_modules_ghost_job-manager: {}
   node_modules_ghost_link-redirects: {}
