+++ ghost/core/package.json
@@ -98,7 +98,6 @@
     "@tryghost/identity-token-service": "0.0.0",
     "@tryghost/image-transform": "1.3.0",
     "@tryghost/importer-handler-content-files": "0.0.0",
-    "@tryghost/importer-revue": "0.0.0",
     "@tryghost/job-manager": "0.0.0",
     "@tryghost/kg-card-factory": "5.1.0",
     "@tryghost/kg-converters": "1.1.0",
@@ -216,11 +215,13 @@
     "nconf": "0.12.1",
     "node-fetch": "2.7.0",
     "node-jose": "2.2.0",
+    "papaparse": "5.3.2",
     "path-match": "1.2.4",
     "probe-image-size": "7.2.3",
     "rss": "1.2.2",
     "sanitize-html": "2.14.0",
     "semver": "7.7.1",
+    "simple-dom": "1.4.0",
     "stoppable": "1.1.0",
     "superagent": "5.1.0",
     "superagent-throttle": "1.0.1",
