--- ghost/importer-revue/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/importer-revue');
